+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - Add warning related to non-unique transaction ids.
 - Improved error handling and parseability with several new error types (see ntropy_sdk.errors), for both client side and server side errors
-- Added `intermediary` to the EnrichedTransaction class containing intermediary merchant info (i.e., payment processor)
+- Added `intermediaries` to the EnrichedTransaction class containing the information for transaction intermediaries (i.e., payment processor)
 
 ## [4.11.0]  2020-10-13
 - Added `ntropy_sdk.recurring_payments` package with support for recurring payments feature